Convert flags to strings when storing
This fixes storage issues when adding values to the flags.  Also attempts to convert all of the values that are stored in a list to a string before they are sent across (since sleekxmpp doesn't handle this gracefully).

Transmit all of the flags across as the text-multi so that they can have multiple values.

Allow for multiple values to be passed to the append method of the results collection payload.